refactor(): Rank를 domain package로 이동
