+++ README.md
@@ -53,8 +53,6 @@ utils.Calculator
 - 구입금액과 당첨금액을 바탕으로 수익률을 계산한다.
 
 ##TODO
-3. Rank를 도메인으로 수정?
-
 ##DONE
 1. 구입금액(LottoAmount)을 입력 받는다.
 2. 로또 금액에 따른 로또 구입 갯수를 반환한다.
@@ -87,6 +85,7 @@ utils.Calculator
 24. Rank의 돈 단위를 2_000_000_000로 관리
 25. Rank에 대한 설명을 view 표시하도록 개선
 26. Uilts성 클래스 삭제
+27. Rank를 도메인으로 수정?
 
 
 ## 피드백 1
