+++ src/main/java/lotto/domain/LottoGame.java
@@ -1,7 +1,6 @@
 package lotto.domain;
 
 import lotto.constant.Lotto;
-import lotto.constant.Rank;
 
 import java.util.List;
 import java.util.stream.Collectors;
