+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -1,12 +1,9 @@
 package lotto.domain;
 
-import lotto.constant.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
